requirements.txt
fastapi
uvicorn[standard]
databases[postgresql]
sqlalchemy
python-multipart
python-dotenv
line-bot-sdk
psycopg2-binary  # 用於 PostgreSQL 連線